Define value sets w/ all codes from code system
In a few cases, the definitions tried to bind fields to a code system, rather than a value set.  This caused import errors.  Now we define value sets containing all the codes from a given system, and then bind the field to that value set (instead of the code system).