Don't put card constraint on numerator/denominator units
Previously, the `Numerator` and `Denominator` constraints were:
```
1..1 Numerator.Quantity with units UCUM#1
1..1 Denominator.Quantity.Units with Coding from http://hl7.org/fhir/ValueSet/units-of-time
```

The problem with that representation is that the `1..1` constraint gets applied to the `Quantity` and `QuantityUnits`, which isn't strictly necessary and causes other problems down the road (since `Quantity.Units` does not map cleanly to a single FHIR element).

Removing the extra `1..1` constraints on the nested paths fixes the problem.