Narrowed the difference between the two branches
There are now only three files that differentiate the
may-2018-ballot-topic-context and may-2018- ballot branches:
cimi_statement.txt, cimi_statement_map.txt, and oncology.txt. This means
anything we do in the other files should be the same on both branches,
in the future. In particular, if we update cimi.txt (core cimi) then it
should be kept in synch as long as these two branches exist.