Revert TumorDimensions to align with BrCa ballot
Reverted TumorDimension1 to TumorLongestDimension, since it provides a more accurate description of the dimension of most interest for AJCC, WHO, and CAP. Revised TumorDimension2 and TumorDimension3 definitions to be more generic and aligned with the SCT codes used. An additional two tumor dimensions are optional in pathology reports, so keeping TumorDimension3 despite not having been part of the BrCa models balloted in September 2018.