Change from Patient to SubjectOfRecord
Discovered cases where Patient was not appropriate attribute name, specifically, when the subject could be a patient, group, location, organization, etc. Introduced the term SubjectOfRecord, which is the Analysis Normal Form term. In ANF this always is a patient, but the way FHIR works, the SoR can be a location or organization or group. This stands in contrast with SubjectOfInformation, which is needed only when an observation deals with someone or something other than the patient. For exmaple, the SubjectOfRecord could be an Organization, and the SubjectOfInformation could be a Location of that organization.